Adding GuessCorrectness + bug fix
checks if the answer given with a guess question is correct. if it is not the case the box will become red and next to it it will give the correct answer.
BUG fix: if a question came around for a second time a new question would be chosen but the old one would show but would be graded by the new question correctness. Is now fixed